Bug fixes, cleanup

+++ my-strava-domain/src/main/java/com/tpo/fitness/domain/activity/Activity.java
@@ -20,7 +20,7 @@ public class Activity {
     private Date startDate;
     private Boolean commute;
     private Boolean manual;
-    private Float calories;
+    private Long calories;
 
     @Override
     public String toString() {
