+++ my-strava-persistence/src/main/java/com/tpo/strava/persistence/entities/ActivityEntity.java
@@ -22,7 +22,7 @@ public class ActivityEntity {
     private String athleteId;
 
     @Column
-    private Float calories;
+    private Long calories;
 
     @Column
     private Float distance;
